+++ alidns-20150109/alibabacloud_alidns20150109/models/_create_ati_agent_register_info_request.py
@@ -13,8 +13,10 @@ class CreateAtiAgentRegisterInfoRequest(DaraModel):
         agent_description: str = None,
         agent_display_name: str = None,
         agent_host: str = None,
+        agent_sub_host: str = None,
         agent_version: str = None,
         client_token: str = None,
+        domain_mode: str = None,
         endpoints: List[main_models.CreateAtiAgentRegisterInfoRequestEndpoints] = None,
         registrant_id: str = None,
     ):
@@ -28,12 +30,14 @@ class CreateAtiAgentRegisterInfoRequest(DaraModel):
         # 
         # This parameter is required.
         self.agent_host = agent_host
+        self.agent_sub_host = agent_sub_host
         # The version of the agent.
         # 
         # This parameter is required.
         self.agent_version = agent_version
         # Provides idempotency. Within 3 minutes, the same value takes effect only once.
         self.client_token = client_token
+        self.domain_mode = domain_mode
         # The endpoint information of the agent.
         # 
         # This parameter is required.
@@ -63,11 +67,17 @@ class CreateAtiAgentRegisterInfoRequest(DaraModel):
         if self.agent_host is not None:
             result['AgentHost'] = self.agent_host
 
+        if self.agent_sub_host is not None:
+            result['AgentSubHost'] = self.agent_sub_host
+
         if self.agent_version is not None:
             result['AgentVersion'] = self.agent_version
 
         if self.client_token is not None:
             result['ClientToken'] = self.client_token
+
+        if self.domain_mode is not None:
+            result['DomainMode'] = self.domain_mode
 
         result['Endpoints'] = []
         if self.endpoints is not None:
@@ -90,11 +100,17 @@ class CreateAtiAgentRegisterInfoRequest(DaraModel):
         if m.get('AgentHost') is not None:
             self.agent_host = m.get('AgentHost')
 
+        if m.get('AgentSubHost') is not None:
+            self.agent_sub_host = m.get('AgentSubHost')
+
         if m.get('AgentVersion') is not None:
             self.agent_version = m.get('AgentVersion')
 
         if m.get('ClientToken') is not None:
             self.client_token = m.get('ClientToken')
+
+        if m.get('DomainMode') is not None:
+            self.domain_mode = m.get('DomainMode')
 
         self.endpoints = []
         if m.get('Endpoints') is not None: