+++ alidns-20150109/alibabacloud_alidns20150109/models/_delete_ati_agent_register_info_request.py
@@ -16,7 +16,7 @@ class DeleteAtiAgentRegisterInfoRequest(DaraModel):
         self.agent_register_info_id = agent_register_info_id
         # Ensures the idempotency of the request. Generate a parameter value from your client to ensure uniqueness across different requests. ClientToken supports only ASCII characters and cannot exceed 64 characters in length.
         # 
-        # > If you do not specify this parameter, the system automatically uses the RequestId of the API request as the ClientToken. The RequestId may be different for each API request.
+        # > If you do not specify this parameter, the system automatically uses the RequestId of the API request as the ClientToken. The RequestId may vary for each API request.
         self.client_token = client_token
 
     def validate(self):