+++ alidns-20150109/alibabacloud_alidns20150109/models/_delete_ati_agent_register_info_response_body.py
@@ -16,7 +16,7 @@ class DeleteAtiAgentRegisterInfoResponseBody(DaraModel):
         self.access_denied_detail = access_denied_detail
         # RequestId
         self.request_id = request_id
-        # Indicates whether the operation was successful. Valid values:
+        # Indicates whether the operation was successful:
         # 
         # - true: The operation was successful.
         #