+++ alidns-20150109/alibabacloud_alidns20150109/models/_describe_ati_agent_register_info_market_request.py
@@ -12,12 +12,19 @@ class DescribeAtiAgentRegisterInfoMarketRequest(DaraModel):
         client_token: str = None,
         max_results: int = None,
         next_token: str = None,
+        trust_level: str = None,
     ):
+        # The endpoint domain name through which the agent provides services externally.
         self.agent_host = agent_host
+        # The agent version.
         self.agent_version = agent_version
+        # The client token that is used to ensure the idempotence of the request. You can use the client to generate the token, but you must make sure that the token is unique among different requests. The token can contain only ASCII characters and cannot exceed 64 characters in length.
         self.client_token = client_token
+        # The maximum number of entries to return in this request.
         self.max_results = max_results
+        # The pagination token for the next query.
         self.next_token = next_token
+        self.trust_level = trust_level
 
     def validate(self):
         pass
@@ -42,6 +49,9 @@ class DescribeAtiAgentRegisterInfoMarketRequest(DaraModel):
         if self.next_token is not None:
             result['NextToken'] = self.next_token
 
+        if self.trust_level is not None:
+            result['TrustLevel'] = self.trust_level
+
         return result
 
     def from_map(self, m: dict = None):
@@ -60,6 +70,9 @@ class DescribeAtiAgentRegisterInfoMarketRequest(DaraModel):
 
         if m.get('NextToken') is not None:
             self.next_token = m.get('NextToken')
+
+        if m.get('TrustLevel') is not None:
+            self.trust_level = m.get('TrustLevel')
 
         return self