+++ alidns-20150109/alibabacloud_alidns20150109/models/_describe_ati_agent_register_info_request.py
@@ -14,7 +14,7 @@ class DescribeAtiAgentRegisterInfoRequest(DaraModel):
         # 
         # This parameter is required.
         self.agent_register_info_id = agent_register_info_id
-        # The client token that is used to ensure the idempotency of the request. You can use the client to generate the token, but you must make sure that the token is unique among different requests. The token can contain only ASCII characters. If you do not specify this parameter, the system automatically uses the value of RequestId as the value of ClientToken. The value of RequestId for each API request may be different.
+        # The client token that is used to ensure the idempotence of the request. You can use the client to generate the token, but you must make sure that the token is unique among different requests. The token can contain only ASCII characters. If you do not specify this parameter, the system automatically uses the value of RequestId as the value of ClientToken. The value of RequestId may be different for each request.
         self.client_token = client_token
 
     def validate(self):