+++ alidns-20150109/alibabacloud_alidns20150109/models/_describe_ati_registrant_request.py
@@ -10,9 +10,9 @@ class DescribeAtiRegistrantRequest(DaraModel):
         client_token: str = None,
         registrant_id: str = None,
     ):
-        # Ensures the idempotence of the request. Generate a parameter value from your client to ensure uniqueness across different requests. ClientToken supports only ASCII characters and cannot exceed 64 characters in length.
+        # Ensures the idempotency of the request. Generate a parameter value from your client to ensure uniqueness across different requests. ClientToken supports only ASCII characters and cannot exceed 64 characters.
         # 
-        # > If you do not specify this parameter, the system automatically uses the RequestId of the API request as the ClientToken. The RequestId may be different for each API request.
+        # > If you do not specify this parameter, the system automatically uses the RequestId of the API request as the ClientToken. The RequestId may differ for each API request.
         self.client_token = client_token
         # The ID of the real-name verified registrant.
         #