+++ alidns-20150109/alibabacloud_alidns20150109/models/_update_ati_alert_settings_request.py
@@ -11,8 +11,13 @@ class UpdateAtiAlertSettingsRequest(DaraModel):
         alert_group: str = None,
         client_token: str = None,
     ):
+        # The list of alert configurations.
         self.alert_config = alert_config
+        # The list of alert notification groups.
         self.alert_group = alert_group
+        # The client token that is used to ensure the idempotence of the request. You can use the client to generate the token, but you must make sure that the token is unique among different requests. The token can contain only ASCII characters and cannot exceed 64 characters in length.
+        # 
+        # If you do not specify this parameter, the system automatically uses the RequestId of the API request as the ClientToken. The RequestId may be different for each API request.
         self.client_token = client_token
 
     def validate(self):